### 2018-06-30 #798 @whedon
[PRE REVIEW]: proEQUIB - IDL Library for Plasma Diagnostics and Abundance Analysis
**Submitting author:** @danehkar (<a href="http://orcid.org/0000-0003-4552-5997">Ashkbiz Danehkar</a>)
**Repository:** <a href="https://github.com/equib/proEQUIB" target ="_blank">https://github.com/equib/proEQUIB</a>
**Version:** v0.0.6
