### 2018-07-02 db2213e
Update paper
