+++ paper/paper.bib
@@ -92,9 +92,8 @@ archivePrefix = "arXiv",
 @TECHREPORT{Howarth:1981,
    author = {{Howarth}, I.~D. and {Adams}, S.},
     title = "{Program EQUIB}",
-booktitle = {University of London},
+   howpublished = {University College London},
      year = 1981,
-   editor = {{Howarth}, I.~D. and {Adams}, S.},
     month = aug,
    adsurl = {http://adsabs.harvard.edu/abs/1981ucl..rept.....H},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
@@ -105,7 +104,7 @@ booktitle = {University of London},
 	{Liu}, X.-W. and {Pritchet}, C.~J. and {Ercolano}, B.},
     title = "{EQUIB: Atomic level populations and line emissivities calculator}",
  keywords = {Software},
-howpublished = {Astrophysics Source Code Library},
+howpublished = {Astrophysics Source Code Library, ascl:1603.005},
      year = 2016,
     month = mar,
 archivePrefix = "ascl",
