### 2018-08-04 #798 @arfon
@mgalloy @scivision @mankoff - would one or more of you be willing to review this submission for JOSS?
