### 2018-08-08 #798 @danehkar
Your list of joss reviewers also shows that @manodeep @mattpitkin @zingale @ziotom78 @benjaminrose @Cadair @nespinoza are comfortable with reviewing topic areas of astronomy and astrophysics.
