### 2018-08-09 2d9b6cf
Update paper.md
