### 2018-08-09 aa00410
Update paper.md
