### 2018-08-09 #798 @danehkar
@arfon I corrected minor typo mistakes. Please regenerate the article PDF after assigning 2 reviewers.
