### 2018-08-27 #899 @mdpiper
I created https://github.com/equib/proEQUIB/pull/1 to add a short note on installing the proEQUIB library. This addresses item 2 in the Documentation checklist.
