### 2018-08-27 equib/proEQUIB#2 @mdpiper
Add information for contributors
See description in https://github.com/atomneb/AtomNeb-idl/issues/1.
