### 2018-08-27 equib/proEQUIB#3 @mdpiper
Add unit tests
See description in https://github.com/atomneb/AtomNeb-idl/issues/2.
