### 2018-08-27 equib/proEQUIB#4 @mdpiper
Create API documentation
See description in https://github.com/atomneb/AtomNeb-idl/issues/4.
