### 2018-09-14 equib/proEQUIB#5 @mdpiper
Move equib routines to externals directory
The `equib_*` routines in the `pro` directory appear to be written by others. To separate them from the work of @danehkar, I'd suggest that they be moved to a new directory under `externals`.
