### 2018-09-14 #899 @mdpiper
My review of **proEQUIB** is complete. The library and its associated paper are appropriate for publication in JOSS, pending @danehkar's responses to the issues I've opened in the repository. Further:

1. https://github.com/equib/proEQUIB/issues/3 is especially important. There's a fair amount of code in the library (including a common block), and no tests. A set of tests that cover the library will increase my confidence that the library is working as it should.
1. The paper reads well, but it could use one more editing pass. There are a couple awkward phrasings that could be smoothed; for example, "...allow to..." in the first and second bullets.

@arfon I'll complete the open items in the review checklist once the issues above have been addressed.
