### 2018-09-15 #899 @arfon
> @arfon I'll complete the open items in the review checklist once the issues above have been addressed.

Excellent news, thanks @mdpiper! @danehkar - please let us know here when you've had a chance to incorporate this feedback.
