### 2018-09-24 #899 @mgalloy
I have checked off as many items on the Review Checklist as I can right now. Remaining items need to be addressed by @danehkar before I can check them off.
