### 2018-11-04 #899 @mgalloy
As mentioned above, I have reviewed everything on the Review Checklist I can right now. The issues opened by @mdpiper need to be addressed by @danehkar to continue.
