### 2018-11-24 3a67426
Note on Installation and minimum IDL version
Add note on Installation and minimum IDL version suggested by @mdpiper and @mgalloy
