### 2018-11-25 866974e
CONTRIBUTING.md added
CONTRIBUTING.md containing information for contributors added to the package, suggested by @mdpiper
