### 2018-11-27 #899 @danehkar
> @mdpiper The `equib_*` routines in the `pro` directory don't appear to be written by the author. To separate them from the author's work, I suggest that they be moved to a new directory under `externals`. I've raised this as [equib/proEQUIB#5](https://github.com/equib/proEQUIB/issues/5).

I moved them to to a `misc` under `externals`, and I renamed  `equib_*` to `_*`.  These routines (interp2d.pro, meshgrid.pro, sign.pro, str2int.pro, strnumber.pro) were chosen from other IDL packages to replace some FORTRAN functions of [EQUIB](https://github.com/equib/EQUIB) during my conversion from FORTRAN to IDL. I added `equib_*` to them to separate from other IDL functions with similar names in other IDL packages in my machine.  Now I closed the issue [equib/proEQUIB#5](https://github.com/equib/proEQUIB/issues/5).
