### 2018-11-27 #899 @danehkar
> @mdpiper I added three issues:
>
> * [equib/proEQUIB#2](https://github.com/equib/proEQUIB/issues/2)
> * [equib/proEQUIB#4](https://github.com/equib/proEQUIB/issues/4)
>
> to the author's repo to address the last three items in the Documentation checklist.

- I added [CONTRIBUTING.md](https://github.com/equib/proEQUIB/blob/master/CONTRIBUTING.md) to the [proEQUIB repository](https://github.com/equib/proEQUIB) on GitHub, which contains some guidelines for contributors. I closed the issue [equib/proEQUIB#2](https://github.com/equib/proEQUIB/issues/2).

- I created API documentation with [IDLdoc](https://github.com/mgalloy/idldoc) and copied them into [docs](https://github.com/equib/proEQUIB/tree/master/docs). You can check it. If there is no problem, I will close the issue [equib/proEQUIB#4](https://github.com/equib/proEQUIB/issues/4).
