### 2018-11-28 #899 @mdpiper
@danehkar Looks good! You can close https://github.com/equib/proEQUIB/issues/4. It would be best to host the documentation at another location, such as rtfd.io or your own site.
