### 2018-11-30 f2ac613
setup appveyor and travis tester
