### 2018-11-30 cc48211
Create .appveyor.yml
