+++ README.rst
@@ -21,26 +21,85 @@ proEQUIB
 Description
 ============
 
-The **proEQUIB** library is a collection of `Interactive Data Language <http://www.harrisgeospatial.com/ProductsandSolutions/GeospatialProducts/IDL.aspx>`_ (IDL)/`GNU Data Language <http://gnudatalanguage.sourceforge.net/>`_ (GDL) programs developed to calculate atomic level populations and line emissivities in statistical equilibrium in multi-level atoms for different physical conditions of the stratification layers where the chemical elements are ionized. This library includes the IDL/GDL implementation of the program `EQUIB <http://adsabs.harvard.edu/abs/2016ascl.soft03005H>`_, which was originally written in FORTRAN by `Howarth & Adams (1981) <http://adsabs.harvard.edu/abs/1981ucl..rept.....H>`_, and was recently converted to IDL/GDL. It also includes the IDL/GDL implementation of deredden functions from `STSDAS IRAF Package <http://www.stsci.edu/institute/software_hardware/stsdas>`_, and atomic data from the `AtomNeb database <https://github.com/atomneb/AtomNeb-idl>`_. It uses the IDL/GDL implementation of the `MIDAS <http://www.eso.org/~ohainaut/ccd/midas.html>`_ scripts by X. W. Liu and codes from `MOCASSIN <https://github.com/mocassin/MOCASSIN-2.0>`_ for some heavy element recombination emissivities. 
+The **proEQUIB** library is a collection of `Interactive Data Language <http://www.harrisgeospatial.com/ProductsandSolutions/GeospatialProducts/IDL.aspx>`_ (IDL)/`GNU Data Language <http://gnudatalanguage.sourceforge.net/>`_ (GDL) programs developed to perform plasma diagnostics and abundance analysis using emission line fluxes measured in ionzed nebulae. It uses the `AtomNeb IDL library <https://github.com/atomneb/AtomNeb-idl>`_ to read collision strengths, transition probabilities, and recombination coefficients for collisionally excited lines (CEL) and recombination coefficients for recombination lines (RL). This IDL package can be used to determine interstellar extinctions, electron temperatures, electron densities, and ionic abundances from the measured fluxes of emission lines. It mainly contains the follwing API functions written purely in IDL/GDL: 
 
-History of codes can be found `here <https://physics.mq.edu.au/~ashkbiz/proequib/history/>`_.
+* **API functions for collisionally excited lines (CEL)** have been developed based on the algorithm of the FORTRAN program `EQUIB <http://adsabs.harvard.edu/abs/2016ascl.soft03005H>`_ written in FORTRAN by `Howarth & Adams (1981) <http://adsabs.harvard.edu/abs/1981ucl..rept.....H>`_. The program EQUIB calculates atomic level populations and line emissivities in statistical equilibrium in multi-level atoms for different physical conditions of the stratification layers where the chemical elements are ionized. Using the IDL/GDL implementation of the program `EQUIB <http://adsabs.harvard.edu/abs/2016ascl.soft03005H>`_, electron temperatures, electron densities, and ionic abundances are determined from the measured fluxes of collisionally excited lines.
 
-Website: `physics.mq.edu.au/~ashkbiz/proequib <https://physics.mq.edu.au/~ashkbiz/proequib/>`_
[…]
