+++ doc/libdata.js
@@ -1,0 +1,188 @@
+/* Index used for searching */
+/*
+   Fields used:
+     url, name, type, filename, authors, routine name, comments, parameters,
+     categories, and attributes
+*/
+title = "API Documentation for proEQUIB";
+subtitle = "IDL Library for Plasma Diagnostics and Abundance Analysis";
+libdata = new Array();
+libdataItem = 0;
+
+
+
+libdata[libdataItem++] = new Array("./calc_abund_c_ii_rl.html", "calc_abund_c_ii_rl.pro", '.pro file in <a href="./dir-overview.html">./ directory</a>', "calc_abund_c_ii_rl.pro", "", "", "", "", "          -1", "");
+  
+  
+  libdata[libdataItem++] = new Array("./calc_abund_c_ii_rl.html#calc_abund_c_ii_rl", "calc_abund_c_ii_rl", 'routine in <a href="./calc_abund_c_ii_rl.html">calc_abund_c_ii_rl.pro</a>', "calc_abund_c_ii_rl.pro", "   Ashkbiz Danehkar   ", "calc_abund_c_ii_rl", "     This function determines the ionic abundance from the observed      flux intensity for the given wavelength of C II recombination line      by using the recombination coefficients from      from Davey et al. (2000) 2000A&AS..142...85D.   ", "temperature                         electron temperature  density                         electron density  wavelength                         Line Wavelength in Angstrom  line_flux                         line flux intensity  c_ii_rc_data                         C II recombination coefficients  h_i_aeff_data                         H I recombination coefficients   ", "Abundance Analysis Recombination Lines", "   0.0.3        Based on recombination coefficients for C II lines from      Davey et al. 2000A&AS..142...85D.       Adopted from MOCASSIN, Ercolano et al. 2005MNRAS.362.1038E.       02/2003, Yong Zhang, added to MOCASSIN.       10/05/2013, A. Danehkar, Translated to IDL code.       15/04/2017, A. Danehkar, Integration with AtomNeb.     This library is released under a GNU General Public License.       For example: <span class= code-prompt >IDL&gt;</span> base_dir = file_dirname(file_dirname((routine_info('$MAIN$', /source)).path)) <span class= code-prompt >IDL&gt;</span> data_rc_dir = ['atomic-data-rc'] <span class= code-prompt >IDL&gt;</span> Atom_RC_All_file= filepath('rc_collection.fits', root_dir=base_dir, subdir=data_rc_dir ) <span class= code-prompt >IDL&gt;</span> Atom_RC_SH95_file= filepath('rc_SH95.fits', root_dir=base_dir, subdir=data_rc_dir ) <span class= code-prompt >IDL&gt;</span> atom='h' <span class= code-prompt >IDL&gt;</span> ion='ii' ; H I <span class= code-prompt >IDL&gt;</span> h_i_rc_data=atomneb_read_aeff_sh95(Atom_RC_SH95_file, atom, ion) <span class= code-prompt >IDL&gt;</span> h_i_aeff_data=h_i_rc_data[0].Aeff <span class= code-prompt >IDL&gt;</span> atom='c' <span class= code-prompt >IDL&gt;</span> ion='iii' ; C II <span class= code-prompt >IDL&gt;</span> c_ii_rc_data=atomneb_read_aeff_collection(Atom_RC_All_file, atom, ion) <span class= code-prompt >IDL&gt;</span> temperature=double(10000.0) <span class= code-prompt >IDL&gt;</span> density=double(5000.0) <span class= code-prompt >IDL&gt;</span> c_ii_6151_flux = 0.028 <span class= code-prompt >IDL&gt;</span> wavelength=6151.43 <span class= code-prompt >IDL&gt;</span> Ab
[…]
