+++ docs/README.md
@@ -1,0 +1,5 @@
+## Documentation
+
+The [API Documentation](https://equib.github.io/proEQUIB/doc/ is available on [equib.github.io/proEQUIB](https://equib.github.io/proEQUIB/).
+
+
