### 2018-12-01 #899 @danehkar
> @mdpiper It would be best to host the documentation at another location, such as rtfd.io or your own site.

The documentation is now hosted at https://equib.github.io/proEQUIB/doc/
