### 2018-12-01 #899 @sundmanbo
Please remove me from this and other JOSS reviewing threads, I have tried
but not managed to do that myself

Bo Sundman

Den lör 1 dec. 2018 kl 08:46 skrev Ashkbiz Danehkar <
notifications@github.com>:

> @mdpiper <https://github.com/mdpiper> It would be best to host the
> documentation at another location, such as rtfd.io or your own site.
>
> The documentation is now hosted at https://equib.github.io/proEQUIB/doc/
>
