### 2018-12-01 #899 @danehkar
I noticed that `triangulate` and `trigrid` in GDL have some bugs, and not work correctly in [interp2d.pro](https://github.com/equib/proEQUIB/blob/master/externals/misc/_interp2d.pro). I replaced [interp2d.pro](https://github.com/equib/proEQUIB/blob/master/externals/misc/_interp2d.pro) with [interp_2d.pro](https://github.com/equib/proEQUIB/blob/master/externals/misc/_interp_2d.pro). Now proEQUIB functions also work in GDL as seen in [Travis CI](https://travis-ci.org/equib/proEQUIB) and [Appveyor](https://ci.appveyor.com/project/danehkar/proequib). To check, you need to install GDL (`sudo dnf install gdl` in Fedora, `sudo apt-get install gnudatalanguage` in Ubuntu, `brew install gnudatalanguage` in OS X).
