### 2018-12-01 2ebd54d
Setup mgunit for automatic test
