### 2018-12-01 #899 @mgalloy
@sundmanbo Unwatch this repo. Go to any page on for the JOSS reviews repo (such as the page for this thread), go to the top of the page, click "Unwatch".
