### 2018-12-01 #899 @arfon
@sundmanbo - yes, please do this:

![](https://cloud.githubusercontent.com/assets/4483/20250593/64d7ce48-a9de-11e6-9225-d3dfb3e48e68.png)
