### 2018-12-02 #899 @danehkar
> I added three issues:
>
> * [equib/proEQUIB#2](https://github.com/equib/proEQUIB/issues/2)
> * [equib/proEQUIB#3](https://github.com/equib/proEQUIB/issues/3)
> * [equib/proEQUIB#4](https://github.com/equib/proEQUIB/issues/4)
>
> to the author's repo to address the last three items in the Documentation checklist.

@mdpiper I included  [mgunit](https://github.com/mgalloy/mgunit) in the folder  [test_mgunit](https://github.com/equib/proEQUIB/tree/master/test_mgunit) for automated tests. You need to have atomic data in ``externals/atomneb/atomic-data`` and ``externals/atomneb/atomic-data-rc`` in do this test. To have all dependencies, you can clone the whole project with dependent libraries:

    git clone --recursive https://github.com/equib/proEQUIB.git

Then, you can run the test for mgunit:

    idl test_all.pro

There should not be any failure in ``test-results.log`` and ``test-results.html`` files.

I closed the issues [equib/proEQUIB#2](https://github.com/equib/proEQUIB/issues/2) and [equib/proEQUIB#4](https://github.com/equib/proEQUIB/issues/4). If there is no problem in the above test, I will close the issue [equib/proEQUIB#3](https://github.com/equib/proEQUIB/issues/3).
