### 2018-12-02 #899 @danehkar
> I have marked all items on the checklist. My review is complete, good luck @danehkar!

@mgalloy Thank you for your test. I much appreciate your review. I am now checking the paper if there is any typos there.
