### 2018-12-02 7ae4f7f
Add BIng verification
