+++ README.md
@@ -1,0 +1,4 @@
+## Documentation
+
+The [API Documentation](https://equib.github.io/proEQUIB/doc/) is available on [equib.github.io/proEQUIB](https://equib.github.io/proEQUIB/). The documentation for proEQUIB is produced by ``proequib_build_doc.pro`` which uses [IDLdoc](https://github.com/mgalloy/idldoc). 
+
