### 2018-12-02 88e5265
Update paper.md
