+++ paper/paper.md
@@ -1,64 +1,71 @@
 ---
 title: 'proEQUIB: IDL Library for Plasma Diagnostics and Abundance Analysis'
 tags:
-  - IDL
+  - astrophysics
+  - gaseous nebulae
   - plasma diagnostics
   - abundance analysis
-  - ionized nebulae
-  - astrophysics
+  - IDL
+  - GDL
 authors:
   - name: Ashkbiz Danehkar
     orcid: 0000-0003-4552-5997
     affiliation: "1, 2"
 affiliations:
- - name: Department of Physics and Astronomy, Macquarie University, Sydney, NSW 2109, Australia
+ - name: Research Centre in Astronomy, Astrophysics and Astrophotonics, Macquarie University, Sydney, NSW 2109, Australia
    index: 1
  - name: Harvard-Smithsonian Center for Astrophysics, 60 Garden Street, Cambridge, MA 02138, USA 
    index: 2
-date: 29 June 2018
+date: 30 June 2018
 bibliography: paper.bib
 ---
 
 # Summary
 
-The emission line fluxes measured from the spectra of ionized nebulae provide 
-insights into their physical conditions and elemental abundances, as well as 
-interstellar extinctions. We can determine temperatures, densities, and chemical 
-abundances from the measured fluxes of _collisionally excited lines_ (CEL) and 
-_optical recombination lines_ (ORL) of ionized nebulae 
+The emission lines emitted from gaseous nebulae carry valuable
+information about the physical conditions and chemical abundances of ionized gases 
+in these objects, as well as the interstellar reddening. We determine the electron temperature, 
+the electron density, and the ionic abundances from the dereddened fluxes of _collisionally excited lines_ (CEL) 
+and _recombination lines_ (RL) identified in nebular spectra 
 [see e.g. @Danehkar:2013; @Danehkar:2014; @Danehkar:2016; @Danehkar:2018].
  
-``proEQUIB`` is a library including several functions developed 
-in the Interactive Data Language (IDL), which can be used to derive 
-electron temperatures, electron densities, and ionic abundances from the measured fluxes. 
-It employs the IDL library ``AtomNeb`` _Atomic Data for Ionized Nebulae_ [@Danehkar:2018b], 
-which contains collision strengths, transition probabilities, and recombination 
-coefficients for collisional excitation and recombination calculations. 
-It includes several IDL functions to determine physical conditions and 
-chemical abundances from CEL and ORL, and estimate interstellar extinctions 
-from Balmer lines.
+``proEQUIB`` is a library including several application programming interface (API) functions developed 
+in the Interactive Data Language (IDL), 
+which can be used to determine temperatures, 
+densities, and chemical abundances from emission lines of ionized nebulae. 
+This IDL library can also be used with the GNU Data Language (GDL) [@Arabas:2010; @Coulais:2010], 
+which is a free and open-source alternative IDL compiler. 
+This IDL/GDL package employs the IDL library ``AtomNeb`` _Atomic Data for Ionized Nebulae_ [@Danehkar:2018b], 
+which contains collision strengths and transition probabilities for collisional excitation calculations, 
+and recombination coefficients for recombination calculations. 
+This package includes several API functions to determine physical conditions and 
+chemical abundances from CEL and RL, derive interstellar extinctions 
+from Balmer lines, and deredden the observed fluxes:
 
-- The CEL functions, which were completely developed in the IDL programming 
-language based on the algorithm of the FORTRAN program EQUIB [@Howarth:1981; @Howarth:2016], 
-allow to determine electron temperatures, electron densities, and ionic abundances 
-from the measured fluxes of _collisionally excited lines_.
[…]
