+++ paper/paper.bib
@@ -2,13 +2,12 @@
 @ARTICLE{Danehkar:2013,
    author = {{Danehkar}, A. and {Parker}, Q.~A. and {Ercolano}, B.},
     title = "{Observations and three-dimensional ionization structure of the planetary nebula SuWt 2}",
-  journal = {\mnras},
+  journal = {Monthly Notices of the Royal Astronomical Society},
 archivePrefix = "arXiv",
    eprint = {1307.2974},
  primaryClass = "astro-ph.SR",
  keywords = {ISM: abundances, planetary nebulae: individual: PN SuWt 2},
      year = 2013,
-    month = sep,
    volume = 434,
     pages = {1513-1530},
       doi = {10.1093/mnras/stt1116},
@@ -20,13 +19,12 @@ archivePrefix = "arXiv",
    author = {{Danehkar}, A. and {Todt}, H. and {Ercolano}, B. and {Kniazev}, A.~Y.
 	},
     title = "{Observations and three-dimensional photoionization modelling of the Wolf-Rayet planetary nebula Abell 48}",
-  journal = {\mnras},
+  journal = {Monthly Notices of the Royal Astronomical Society},
 archivePrefix = "arXiv",
    eprint = {1403.0567},
  primaryClass = "astro-ph.SR",
  keywords = {stars: Wolf-Rayet, ISM: abundances, planetary nebulae: individual: Abell 48},
      year = 2014,
-    month = apr,
    volume = 439,
     pages = {3605-3615},
       doi = {10.1093/mnras/stu203},
@@ -40,7 +38,6 @@ archivePrefix = "arXiv",
  keywords = {planetary nebulae, Wolf-Rayet stars, Interstellar medium, kinematics, abundances, Stellar evolution,  Imaging Spectroscopy},
    school = {Macquarie University, Australia},
      year = 2014,
-    month = April,
       doi = {10.5281/zenodo.47794},
    adsurl = {http://adsabs.harvard.edu/abs/2014PhDT........76D},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
@@ -49,13 +46,12 @@ archivePrefix = "arXiv",
 @ARTICLE{Danehkar:2016,
    author = {{Danehkar}, A. and {Parker}, Q.~A. and {Steffen}, W.},
     title = "{Fast, Low-ionization Emission Regions of the Planetary Nebula M2-42}",
-  journal = {\aj},
+  journal = {The Astronomical Journal},
 archivePrefix = "arXiv",
    eprint = {1601.01702},
  primaryClass = "astro-ph.SR",
  keywords = {ISM: jets and outflows, planetary nebulae: individual: M2-42, stars: evolution},
      year = 2016,
-    month = feb,
    volume = 151,
       eid = {38},
     pages = {38},
@@ -67,13 +63,12 @@ archivePrefix = "arXiv",
 @ARTICLE{Danehkar:2018,
    author = {{Danehkar}, A.},
     title = "{Bi-Abundance Ionisation Structure of the Wolf-Rayet Planetary Nebula PB 8}",
-  journal = {\pasa},
+  journal = {Publications of the Astronomical Society of Australia},
 archivePrefix = "arXiv",
    eprint = {1801.00892},
  primaryClass = "astro-ph.SR",
  keywords = {ISM: abundances, planetary nebulae: individual (PB 8), stars: Wolf-Rayet},
      year = 2018,
-    month = jan,
    volume = 35,
       eid = {e005},
     pages = {e005},
@@ -82,11 +77,12 @@ archivePrefix = "arXiv",
[…]
