### 2018-12-02 #899 @danehkar
@arfon I uploaded the updated paper into the folder [`paper`](https://github.com/equib/proEQUIB/tree/master/paper). Previously, there were some issues in ``paper.bib`` and the paper pdf file did not show the journal names.
