### 2018-12-02 #899 @danehkar
@arfon The paper looks fine, but there is an issue in the citation order and style:

- For `[see e.g. @Danehkar:2013; @Danehkar:2014; @Danehkar:2016; @Danehkar:2018]`:

  > (Danehkar, 2018a; see e.g. Danehkar, Parker, & Ercolano, 2013; Danehkar, Parker, & Steffen, 2016; Danehkar, Todt, Ercolano, & Kniazev, 2014)

  should be:

  > (see e.g. Danehkar, Parker, and Ercolano 2013; Danehkar et al. 2014; Danehkar, Parker, and Steffen 2016; Danehkar 2018a)

- For `[@Ercolano:2003; @Ercolano:2005]`:

  > (Ercolano, Barlow, & Storey, 2005; Ercolano, Barlow, Storey, & Liu, 2003)

  should be

  > (Ercolano et al. 2003; Ercolano, Barlow, and Storey 2005)

- For `[@Danehkar:2016; @Danehkar:2018]`:

  > (Danehkar, 2018a; Danehkar et al., 2016)

  should be

  > (Danehkar, Parker, and Steffen 2016; Danehkar 2018a)

- For `[see e.g. @Danehkar:2013; @Danehkar:2014; @Danehkar:2014b]`:

  > (Danehkar, 2014; see e.g. Danehkar et al., 2013, 2014)

  should be

  > (see e.g. Danehkar, Parker, and Ercolano 2013; Danehkar et al. 2014; Danehkar 2014)

- I do not think it is necessary to show more than 3 authors in text:

  > (Arabas, Schellens, Coulais, Gales, & Messmer, 2010; Coulais et al., 2010)

  should be

  > (Arabas, et al., 2010; Coulais et al., 2010)

The citation style in the previous PDF file did not have any problems.
