### 2018-12-03 b9edcfe
Update paper.md
Changed to the APA citation style.
