+++ paper/paper.md
@@ -27,46 +27,49 @@ information about the physical conditions and chemical abundances of ionized gas
 in these objects, as well as the interstellar reddening. We determine the electron temperature, 
 the electron density, and the ionic abundances from the dereddened fluxes of _collisionally excited lines_ (CEL) 
 and _recombination lines_ (RL) identified in nebular spectra 
-[see e.g. @Danehkar:2013; @Danehkar:2014; @Danehkar:2016; @Danehkar:2018].
+(see e.g. @Danehkar:2013; @Danehkar:2014; @Danehkar:2016; @Danehkar:2018).
  
 ``proEQUIB`` is a library including several application programming interface (API) functions developed 
 in the Interactive Data Language (IDL), 
 which can be used to determine temperatures, 
 densities, and chemical abundances from emission lines of ionized nebulae. 
-This IDL library can also be used with the GNU Data Language (GDL) [@Arabas:2010; @Coulais:2010], 
+This IDL library can also be used by the GNU Data Language (GDL) (@Arabas:2010; @Coulais:2010), 
 which is a free and open-source alternative IDL compiler. 
-This IDL/GDL package employs the IDL library ``AtomNeb`` _Atomic Data for Ionized Nebulae_ [@Danehkar:2018b], 
+This IDL/GDL package employs the IDL library ``AtomNeb`` _Atomic Data for Ionized Nebulae_ (@Danehkar:2018b), 
 which contains collision strengths and transition probabilities for collisional excitation calculations, 
 and recombination coefficients for recombination calculations. 
 This package includes several API functions to determine physical conditions and 
 chemical abundances from CEL and RL, derive interstellar extinctions 
 from Balmer lines, and deredden the observed fluxes:
 
-- The API functions for _CEL analysis_ were developed in the IDL programming 
-language based on the algorithm of the FORTRAN program EQUIB [@Howarth:1981; @Howarth:2016]. 
+- The API functions for the _CEL analysis_ were developed in the IDL programming 
+language based on the algorithm of the FORTRAN program ``EQUIB`` (@Howarth:1981; @Howarth:2016), 
+which calculates atomic level populations and line emissivities in 
+statistical equilibrium in multi-level atoms for the given physical conditions. 
 These API functions can be used to determine the electron temperature, the electron density, 
 and the ionic abundances from the dereddened fluxes of _collisionally excited lines_ emitted from 
 ionized gaseous nebulae.
 
-- The API functions for _RL analysis_ were developed in IDL according to the algorithm 
+- The API functions for the _RL analysis_ were developed in IDL according to the algorithm 
 of the recombination scripts by X. W. Liu and Y. Zhang included in the FORTRAN 
-program MOCASSIN [@Ercolano:2003; @Ercolano:2005]. These API functions can be used to 
+program ``MOCASSIN`` (@Ercolano:2003; @Ercolano:2005). These API functions can be used to 
 determine the ionic abundances from the dereddened fluxes of _recombination lines_ emitted from 
 ionized nebulae.
 
-- The API functions for _reddening analysis_ were developed based on the methods 
-of the reddening functions in STSDAS IRAF Package [@Bushouse:1994; @Shaw:1994]. 
+- The API functions for the _reddening analysis_ were developed based on the methods 
+of the reddening functions from the Space Telescope Science Data Analysis System 
+(``STSDAS``) IRAF Package (@Bushouse:1994; @Shaw:1994). 
 These API functions can be employed to obtain _interstellar extinctions_ for 
 different reddening laws from the observed fluxes of Balmer lines detected 
[…]
