### 2018-12-03 #899 @danehkar
@arfon I now switched to the APA citation style. I checked it here [#798](https://github.com/openjournals/joss-reviews/issues/798). You can reproduce the article proof. Thanks.
