### 2018-12-03 #899 @mdpiper
@danehkar Thanks for addressing all the issues I raised. I've checked off the last item on my review checklist, so my review is complete!
